Make upgrade workflows more accessible to third parties (#1128)
The`PULUMI_BOT_TOKEN` requirement for third parties is not very
intuitive and the upgrade workflows fail without this available.

1. Add the option to use `PULUMI_PROVIDER_AUTOMATION_TOKEN` instead of
the `PULUMI_BOT_TOKEN` as a more descriptive alternative.
2. Add fallback if no PAT is available to use the built-in GitHub
Actions token. A `permissions` block is therefore also added to allow it
to push commits, create Github issues and pull requests in the
repository.

Using the built-in token has the downside that the created PR will not
be automatically built because it was created by an App ([see the
docs](https://docs.github.com/en/actions/security-for-github-actions/security-guides/automatic-token-authentication#using-the-github_token-in-a-workflow)
for details).

Internally, we should also switch to using
`PULUMI_PROVIDER_AUTOMATION_TOKEN` instead of `PULUMI_BOT_TOKEN` as it's
a better name that describes its purpose better within the ever-more
crowded list of org-level secrets.

Contributes to: #1087