Regenerate workflows for external providers weekly (#1187)
Pulumi managed providers receive Github Actions workflow changes via a
workflow run in `ci-mgmt`, with changes pushed to the provider
repositories.

External providers need a pull based model to take similar workflow
changes in, hence the creation of the `external-provider` template
folder and adding the required workflow in there.

I tested the contents of the workflow standalone first on a single
provider and it worked fine. Here was the generated PR to update the
workflows:

https://github.com/pulumiverse/pulumi-acme/pull/93