Reuse provider build for testing (#1196)
Fixes #967 
Fixes #1033

Unlike the ticket spec in #967 , only linux-amd64 binary is tested by
the full set of integration tests. We might consider opening a follow up
to smoke-test every single one of the platform builds, or else keep #967
open.

As to the ticket spec in #1033 the dependency from build_provider to
prerequisites is real because prerequisites builds the schema and the
schema is needed in the binaries. This needs to stay.

Test run: https://github.com/pulumi/pulumi-azuread/pull/1589

Additional exploration: in prerequisites we run `make test_provider`
which seems to try running unit tests in the `./provider` directory.
Unfortunately some tests in `./provider` directory are also integration
tests, e.g. in pulumi-aws and pulumi-azuread, and the convention of
passing `go test -short` to only run the unit tests does not seem to be
observed universally. For this reason removing `make provider` from
prerequisites is not yet possible as it breaks our builds. This is
likely out of scope of this PR.