Disarm any go:embed directives before running golangci-lint for native providers (#1078)
The pulumi-kubernetes provider embeds a minified version of a Pulumi
JSON schema in the built binary. This file is not checked into source,
only the non minified version is. `golangci-lint` is unable to compile
the provider for linting if the file does not exist. This PR ensures
that we disable any go:embed directives by doing a simple find/replace.

This is necessary as `golangci-lint` is unable to skip processing
files/directories. Instead, the `exclude-dirs/files` directives will
still parse those files/directories, but just suppress warnings/errors.
If it is not compilable, in this case due to missing embed files, the
linter will also fail. See
https://golangci-lint.run/welcome/faq/#why-do-you-have-typecheck-errors

Seehttps://github.com/pulumi/pulumi-kubernetes/pull/3355 for example
passing run that modifies the workflow file directly.