Always emit verbose test output (#1212)
This might be intentional for the sake of easier debugging of failures,
but we're going to start moving some tests around for the sake of
sharding and it will be very important to be able to see the exact tests
being run. Otherwise we run the risk of accidentally dropping something.

Let's at least turn this on temporarily.