Sign windows binaries via makefile (#1227)
Move the logic for signing windows binaries into the makefile from the
CI scripts. This allows testing with signed binaries locally.

- Fail with error if running in CI so we don't release unsigned
binaries. Just warn locally.
- Add inline documentation explaining how to fix the missing
configuration.
- Allow third party providers to skip signing by having none of the
configuration set. Perhaps we could make this more explicit in the
future, but that would potentially require changing the config on a
bunch of providers - one side or the other.