Allow root-level go.mod for setup-go caching (#1226)
It is valid (and
[preferable](https://go.dev/wiki/Modules#should-i-have-multiple-modules-in-a-single-repository))
to have a root-level `go.mod`. If one is present we should use it as a
cache key for setup-go.

Additionally since there is some talk of moving the SDK's module down a
level we might as well handle that case gracefully as well, even if
we're not doing it at the moment.

(I would prefer to use `**/*.sum`, but we don't want Go examples to bust
the cache.)