Set permissions on the calling side. (#1233)
The `permissions` block in a reusable workflow doesn't grant these
permissions. These are more used as an early indication of the
permissions required by the reusable workflow. The correct permissions
still need to be set on the calling side.

See `Reusable workflows` in [this blog
article](https://www.kenmuse.com/blog/github-actions-workflow-permissions/).

Follow up to #1214.

Without setting the `permissions` on the calling side, the reusable
workflow will not have the correct permissions.

The Pulumiverse runs failed without the required `permissions` block:
https://github.com/pulumiverse/pulumi-acme/actions/runs/12365788538
After adding the `permissions` block on the calling side:
https://github.com/pulumiverse/pulumi-acme/actions/runs/12369028130

https://github.com/pulumiverse/pulumi-acme/pull/98

For Pulumiverse, this fails because the organization wide workflow
settings are configured to `Read repository contents and packages
permissions`.