AWS Native: Add AWS credentials before Test Provider Library step (#1239)
In https://github.com/pulumi/pulumi-aws-native/pull/1919 I want to run
an actual end-2-end test with `pulumitest` in AWS Native, which requires
AWS credentials. In understand that a larger consolidation may be coming
with https://github.com/pulumi/ci-mgmt/issues/1101, but this change
would allow me to land my test ahead of that. Please let me know if
there is a better way of achieving the same outcome.