Automatically install pulumictl (#1284)
Bridged provider Makefile used to assume that `pulumictl` is available
in PATH. With this change it will now automatically install it if this
is not the case. This is easier to use and also will help `make
renovate` targets succeed without having `pulumictl` installed.

The version of the tool is controlled so that it is in sync across the
Makefile and CI.