Configure Git to checkout files with long names (#1323)
On windows Action runners, the actions/checkout action needs extra
configuration to be able to handle long file names.

This pull request adds this setting for the release-verification job,
which uses both Linux and Windows runners.

While this was discovered while trying to enact release verification for
pulumi-gcp, as it is just a configuration setting it's probably worth it
to enable nonconditionally (on every provider) so we don't have to
re-discover that we need to twist this knob when we enable release
verification elsewhere.

Fixes https://github.com/pulumi/ci-mgmt/issues/1322