Add windows signing for native provider releases (#1318)
This PR enables Windows singing for native provider releases made with
GoReleaser. These changes require the `make sign-goreleaser-exe` target
to exist. This is done in the individual provider repos as the Makefiles
are not managed centrally by ci-mgmt.

For testing purposes, these changes are also copied to their respective
native provider repos with a prerelease tag cut to validate that the
binaries are signed.

Confirmed native providers with signed prerelease windows builds:

-
[kubernetes](https://github.com/pulumi/pulumi-kubernetes/actions/runs/12996104839)
-
[aws-native](https://github.com/pulumi/pulumi-aws-native/actions/runs/12996105418)
-
[command](https://github.com/pulumi/pulumi-command/actions/runs/12996105991)
-
[docker-build](https://github.com/pulumi/pulumi-docker-build/actions/runs/12996106508)
-
[kubernetes-cert-manager](https://github.com/pulumi/pulumi-kubernetes-cert-manager/actions/runs/12997632596)
-
[kubernetes-coredns](https://github.com/pulumi/pulumi-kubernetes-coredns/actions/runs/12997647427)
-
[kubernetes-ingress-nginx](https://github.com/pulumi/pulumi-kubernetes-ingress-nginx/actions/runs/12997658107)
-
[pulumistack](https://github.com/pulumi/pulumi-pulumistack/actions/runs/12996108738)
-
[google-native](https://github.com/pulumi/pulumi-google-native/actions/runs/12996109430)