Set gcloud auth permissions on each publish step  (#1332)
Followup to https://github.com/pulumi/ci-mgmt/pull/1331 which was
supposed to fix https://github.com/pulumi/pulumi-gcp/issues/2919 but
instead pushed the permissions problem up into the next parent Workflow.
Upshot is that the "invalid Workflow" error happens in a different file:
https://github.com/pulumi/pulumi-gcp/actions/runs/13042448162/workflow#L101

I believe this is all of the places where we need to tell Actions to
inherit slightly elevated permissions.