Put `generate_coverage_data` into a `main-post-build` job (#1361)
We need to extend the main workflow to run a job only for
internal-bridged providers and would like to avoid scattering
conditionals through the code (such as NoUpstream).

We'll add a hook in the main workflow called "post-build" which defaults
to a stub, skipped job. In the `internal-bridged` templates directory,
we'll override this job to perform the existing `generate_coverage_data`
job.

This also solves the problem in pulumiverse where the
`generate_coverage_data` job fails on every merge to the default branch.

Stacked on #1360