ci: use go version from provider to update workflows (#1399)
### Proposed Changes

This PR updates the update-workflows workflow to ensure we use a Golang
version compatible with the respective provider for updating its own
workflows. It achieves this by first checking out the targeted provider
repository and extracting the Golang version specified in its
`.ci-mgmt.yaml` config file. If no version is specified, the workflow
defaults to the latest "stable" Go release.

This change is necessary because our providers use different Go
versions, and mismatches can lead to build failures due to incompatible
toolchain versions.

Manual testing was conducted by dispatching the workflow against
repositories with and without a defined Go version in their config:
- **AWS (with Go version specified):**
[[Logs](https://github.com/pulumi/ci-mgmt/actions/runs/13444914246/job/37567929144)]
- **GCP (without):**
[[Logs](https://github.com/pulumi/ci-mgmt/actions/runs/13444937729/job/37568003062)]

### Relevant Issues (optional)

Closes #1382.