Carry env vars into upgrade-bridge.yml (#1418)
This came up during investigating bridge downstream checks and in
particular:

https://github.com/pulumi/pulumi-mongodbatlas/pull/714

There is an environment variable that affects whether preview resources
are included in the schema or not:


https://github.com/pulumi/pulumi-mongodbatlas/blob/1a78223a600a451687b4a0b6ae76a6e21db06a71/.ci-mgmt.yaml#L7

This environment variable is set in all workflows but not in
upgrade-bridge.yml. Because it is not set there, the wrong schema/sdk is
generated during bridge upgrade.

This PR suggests to set env vars uniformly. 

Alternatively we may consider wiring this env var into the Go code in
mongodbatlas.