Move native providers to templates (#1480)
We'd like to make CI for our native providers more consistent with the
rest of our provider fleet. The first step is to start using the same
template system for these providers.

I've taken the rather tedious, but safe route of creating templates that
produce almost identical results to the current native-provider-ci
(modulo the ordering of a few env vars). I've moved all the templated
workflows for these providers from native-provider-ci/providers to
provider-ci/test-providers to verify. This should make it easier review
and simplifies adding a .ci-mgmt.yaml to each of these providers.

After this lands we can add a .ci-mgmt.yaml to each of these providers,
add them to providers.json, and remove native-provider-ci. Then we'll be
free to start migrating these templates to share more with other
provider templates.

## Provider PRs
- https://github.com/pulumi/pulumi-command/pull/733
- https://github.com/pulumi/pulumi-aws-native/pull/2165
- https://github.com/pulumi/pulumi-docker-build/pull/511