Run integration tests for Pulumi YAML (#1491)
This PR adds Pulumi YAML as a language in the integration test matrix.
Because Pulumi YAML doesn't have SDKs, we skip the `Download SDK`,
`UnTar SDK folder` and `Install dependencies` steps for Pulumi YAML.

The first commit is the manual change. The second commit is the result
of running `make build` to regenerate examples.

Fixes #1487