Don't mention tfgen binaries in native providers names (#1492)
Native providers don't always have a generation binary, and the if they
do, it is never from `tfgen` (Terraform generation). This PR changes the
name of the download job from `Download provider + tfgen binaries` to
`Download provider`. This is purely a cosmetic change.

The first commit is the manual change. The second commit is the result
of running `make build` to regenerate checked in CI scripts.