Add exceptions to support pulumi-terraform (#1493)
This enables ci-mgmt to generate the CI for
https://github.com/pulumi/pulumi-terraform (post
https://github.com/pulumi/pulumi-terraform/pull/769).

The PR does not effect any other providers. To match the CI scripts
found in https://github.com/pulumi/pulumi-terraform/pull/769, this PR
needs to be stacked on top of #1491 and #1492.

dabadc905acabe096d2a62bd8cc50ab5b35ddd0b is a stack of all the PRs
needed to get pulumi-terraform to fully comply with ci-mgmt. Running
`make ci-mgmt` in pulumi-terraform on that version yields only the
changes shown in
<https://github.com/pulumi/pulumi-command/pull/737/files>.