Applying upstream patches should ignore whitespace differences
I ran into this in
[pulumi-gcp](https://github.com/pulumi/pulumi-gcp/pull/3176) where one
of the patches will apply with `git apply` because `apply` ignores
whitespace differences, but the patch fails to apply with `git am` because
`am` doesn't ignore whitspace differences.

This should be safe to take because if the patches are created/applied
on one machine, but upstream was created on another, whitespace changes
may be unavoidable and do not cause a material difference.