Drop wildcard to prevent accidental collisions (#1485)
I’m helping @scraly from OVH to adopt `ci-mgmt`. In their repositories,
they have a top-level folder called `LICENSES` which is populated by a
central process.

https://github.com/ovh/pulumi-ovh

When rolling out the `external-provider-template` for the
`ovh/pulumi-ovh` repository, the copy command in `.make/build_nodejs`
includes `../../LICENSE*` (with wildcard). This wildcard matches folder
`LICENSES` but the copy command fails because it only expects files.

I checked the filenames of the `LICENSE` files in `pulumi` and
`pulumiverse` organization and all seem to be an exact match with
`LICENSE` without file extension.

* https://github.com/search?q=org%3Apulumi%20LICENSE.txt&type=code
* https://github.com/search?q=org%3Apulumi%20LICENSE.md&type=code
*
https://github.com/search?q=org%3Apulumiverse+LICENSE.txt&type=code&p=1
* https://github.com/search?q=org%3Apulumiverse+LICENSE.md&type=code&p=1

So I think this is a safe change.