Remove native-provider-ci (#1510)
All the providers that had been using native-provider-ci to manage their
workflows have been moved to use a template in provider-ci instead.
(google-native is currently unmanaged as it is on the road to
deprecation.)

This means we can delete native-provider-ci and all the workflows &
tools that were pointing to it so we only have one system to maintain
going forward!