Configurable tool versions in native templates (#1520)
Makes the tool versions configurable based on `toolVersions`, affecting
the `GOVERSION` workflow variable (and other languages).

For test purposes, we override the go tool version in the kubernetes
provider, and observe that the `GOVERSION` is set correctly.

We may want to bump the default Go version to 1.24 (in
defaults.config.yaml), preferably in a follow-up.