Add workflow to export repo secrets (#1537)
This repository has repository-specific secrets that need to be migrated
to ESC. These changes add a GitHub Actions workflow to perform this
migration.

The workflow is configured to be run manually. It is only expected that
the worflow will be run once. The workflow uses
https://github.com/pulumi/esc-export-secrets-action to create a new ESC
environment to hold the repository's secrets. This environment also
imports the environment that holds the organization's secrets.