Use pull request base as ref for schema-tools (#1545)
This pull request changes the `default_branch` input to schema-tools to
use the base branch of the pull request that the PR is being made
against.

This means schema-tools can compare changes on a feature branch, or a
stack of PRs, and only show incremental changes, rather than only
comparing against the repository default branch.