Add a github workflow to comment on stale provider bugs (#1633)
We've discussed this a few times, but never quite pulled the trigger.
This picks up the last PR: [[DO NOT MERGE] Create stale_issues.yaml ·
pulumi-aws/5366](https://github.com/pulumi/pulumi-aws/pull/5366) which
tested this workflow against AWS. For this PR, I've kept the logic
identical to that test, just moving it to ci-mgmt and running it on a
cron:

The action is configured to look for kind/bug issues that have not been
updated in the last and have fewer than 2 upvotes. It would comment on
these issues with the text:
> Unfortunately, it looks like this issue hasn't seen any updates in a
while. If you're still experiencing this issue, could you leave a quick
comment to let us know so we can prioritize it?

and add the awaiting-feedback label. If a non-pulumi user replies, the
issue will be bumped back into the triage queue using the usual
automation. If no one replies, these issues will eventually show up in
our "stale awaiting feedback" query on the team dashboards so we can
take one last look at them and decide if we should close them.

The workflow defaults to just doing a dry run and I've added a new
config flag to turn off dryRun and comment/label the issues for real.
(So we can roll this out gradually to provider repos or disable it for a
single repo if needed.)