Add a provider workflow to export repo secrets (#1578)
This adds a new "Export secrets to ESC" workflow to our internal
providers. The
workflow itself is copy-pasted from Pat's prior work. This workflow is
necessary in cases where the provider relies on repo-specific secrets.
It may
need to be run at most once per provider, and we can remove this later
when all
secrets have been migrated.

Confirmed working on docker-build:

```
❯ esc env get pulumi/github-secrets/pulumi-pulumi-docker-build

   Value

    {
      "environmentVariables": {
        ...
      }
    }

   Definition

    imports:
      - imports/github-secrets
    values:
      environmentVariables:
        CODECOV_TOKEN:
          ...
        DOCKER_HUB_PASSWORD:
          ...
```

Refs https://github.com/pulumi/ci-mgmt/issues/1481