Add an ESC shim action (#1588)
This adds a small actions which has the same "interface" as ESC, but
which works with our existing CI environments. The action is trivial and
implemented as vanilla JavaScript to eliminate packaging pitfalls. (Only
JavaScript actions support the dynamic outputs we need.)

The action takes the CI job's environment and makes it available as an
output on the step, in the same way we will eventually consume secrets
as outputs from the esc-action step.

A new template method, `renderESCStep`, is also introduced which allows
us to toggle between the shim action and the actual ESC action. This
will allow us to roll out ESC gradually, and without needing to
introduce ESC conditionals around all of our steps.

This is a no-op. A subsequent PR will actually consume the new action.

Refs https://github.com/pulumi/ci-mgmt/issues/1481