Fix GITHUB_TOKEN permissions for ESC (#1705)
When ESC is enabled our GITHUB_TOKEN needs `id-token: write`
permissions. I added this to each job that uses the ESC step. As a bonus
our steps have less write access than they normally would by default.

As part of this I'm also going to force the boilerplate to enable ESC.
Normally we would enable it in the repo's `.ci-mgmt.yml`, but we have a
chicken-and-egg problem while the ESC-enabled workflows are broken.

Refs https://github.com/pulumi/ci-mgmt/issues/1481.