Don't export all ESC secrets to the environment (#1716)
The default behavior of esc-action is to export everything to the
runner's environment. While this is convenient, it also risks leaking
secrets.

Let's not export these secrets and instead reference them explicitly
where they are needed.

I previously had some config in place for this parameter, but it's
unused so we can remove it.

Refs https://github.com/pulumi/ci-mgmt/issues/1481.