Fix `esc-action` shim generation for external providers (#1736)
The introduction of the ESC secrets support introduced a regression.
External providers which have the ESC support disabled did not have the
`esc-action` shim generated. Also the `terraform-module` test provider
didn't have the shim in its setup.

I refactored the `mise` template into an `all` template for files which
should be generated for all providers without exception. Moving the
`esc-action` shim into `all` resolves the problem for external providers
and `terraform-module`, but has the side-effect that the shim is also
generated for `pulumi-provider-boilerplate`. I did it in this smallest
possible way because restructuring any other way lead me to a rabbit
hole every single time.

---------

Co-authored-by: Matthew Jeffryes <mjeffryes@pulumi.com>