Fix publish external providers (#1739)
Fixes: #1738

In #1736, I changed some code so a set of environment variables was only
added for Pulumi managed providers. In that set of environment
variables, I missed these contain the secrets to publish to the various
package registries.

I tweaked it now that these secrets are always set, but only more are
added for Pulumi managed providers.

Tested manually on `pulumiverse/pulumi-grafana`:

* Before the fix:
https://github.com/pulumiverse/pulumi-grafana/actions/runs/17937299375
* After the fix:
https://github.com/pulumiverse/pulumi-grafana/actions/runs/17968917504