Replace setup-tools with mise (#1730)
This PR replaces the `setup-tools` action with `mise-action`. It also
does a couple of additional structural changes:

1. Moves `mise` configuration to `.config/mise.toml`. This allows users
(specifically external users) to put a `mise.toml` file at the root of
the directory which will be merged on top of the `.config/mise.toml`
file. see https://mise.jdx.dev/configuration.html
2. Adds a migration to delete the old generated `mise.toml` file.
3. Remove the installation of the Pulumi CLI from `plugins.mk` so that
we completely rely on mise.

### Core Benefits

With this setup local development and CI will be the exact same.

- Developers can run `mise install` to install all the tools at the
_exact_ same version as CI. (with the exception of plugins, #1474)
- When something updates the version of a tool in `mise.toml`, `mise
install` will automatically update the lockfile.

### Migrations
I added a couple of migrations to make this rollout easier.

1. Added a .gitignore rule for the `mise.local.toml` file since that is
recommended.
2. Generate a `mise.lock` file if it doesn't exist. This is needed to
workaround [this mise
issue](https://github.com/jdx/mise/discussions/6339). Our migrations can
set the env vars so we don't have to rely on mise to set them. Once the
lockfile is generated with the correct version we can continue to
install the correct version.
4. Migrates the `toolVersions` overrides from `.ci-mgmt.yaml` to a root
level `mise.toml`.

### Determining versions
For the `go` and `pulumi` versions I created a `get-versions.sh` script.
This is partly needed to workaround [this mise
issue](https://github.com/jdx/mise/discussions/6374) which prevents us
from using mise templates to run commands.

- `go`: I am picking the `toolchain` directive version from the `go.mod`
file as the go version to use. This should match the existing logic of
the `install-go` action which automatically uses that version if set.
- `pulumi`: I am using the version of `github.com/pulumi/pulumi/pkg/v3`
from the `go.mod` file.
- In the `test` workflow I am setting the `MISE_ENV` variable to `test`
so that the `mise.test.toml` file will override and use the "latest"
pulumi version. Today we use the `dev` version, but it looks like `mise`
does not know how to install the `dev` version.

I tested this on a couple of providers:
[pulumi-docker](https://github.com/pulumi/pulumi-docker/pull/1519) &
[pulumi-aws](https://github.com/pulumi/pulumi-aws/pull/5832)

fixes #1610