Export the `PULUMI_PROVIDER_AUTOMATION_TOKEN` secret (#1741)
When I aligned with @danielrbradley some months ago on more agnostic
name for an automation token instead of `PULUMI_BOT_TOKEN`, we decided
on `PULUMI_PROVIDER_AUTOMATION_TOKEN`.

All external providers leveraging `ci-mgmt` use this secret name, hence
it needs to be exported as an environment variable for the ESC shim to
function correctly.

The custom token is required for any PR in a provider repo to be created
with a separate identity than `github-actions` so the linked workflows
for a PR get triggered.