Allow providers to upgrade to golangci-lint v2 (#1753)
With #1730 a provider is now able to opt-in to golangci-lint v2 (by
defining an override in mise.toml), but this isn't sufficient to use it.
ci-mgmt still publishes `.golangci.yml` to every provider using the v1
format.

This PR introduces a migration which will check whether the provider is
using v2 and, if it is, the config will be migrated to the expected v2
format. The lint job is also updated to use the setup-mise action in
order to ensure we use a consistent version of `golangci-lint`.

Eventually I would like ci-mgmt's role in linter configuration to be
driven more by this migration instead of a static file. In other words,
instead of blindly dropping a file into every repo we would:
* Check if `.golangci.yml` exists, and, if it doesn't, commit some sane
defaults.
* If `.golangci.yml` exists, ensure it includes _at least_ the expected
defaults.

This would allow providers to opt-in to more stringent (and custom!)
lint rules going forward.

Fixes https://github.com/pulumi/ci-mgmt/issues/1543.
Unblocks https://github.com/pulumi/pulumi-gitlab/issues/980.