Ignore mise.lock in `git-status-check`
The mise.lock file is only updated when a machine first downloads a tool
version. If there is a cache on the machine, the lockfile is not updated
with the checksum entry. It also only adds a checksum entry for the arch
of the machine that is installing the tool. This means that sometimes CI
will update the mise.lock file and cause the git-status-check to fail.