Migrate ci-mgmt `plugins` entries to mise
This PR migrates the `plugins` entry in `ci-mgmt.yaml` to
`.config/mise.toml`. To simplify the mise migrations I also refactored a
little so that we now have the following flow:

- `deleteOldMiseConfig`: Runs first and deletes the old root level
  `mise.toml` if it exists and is not the overrides file
- `migrateCimgmtOverrides`: Migrates things from `ci-mgmt.yaml` to the
  root level override file (currently just `toolVersions`).
- `migrateCimgmtToMise`: Migrates things from `ci-mgmt.yaml` to the
  `.config/mise.toml` file.
- `maintainMiseLock`: Runs last and runs `mise install` to update the
  lockfiles.

I've set the plugins version to "latest" for all the plugins so that we
can implement an upgrade workflow (e.g. via `mise upgrade`). Also, since
the plugins are added to the `.config/mise.toml` we can add specific
version overrides to the root level `mise.toml`.

closes #1474