Migrate ci-mgmt `plugins` entries to mise (#1792)
This PR migrates the `plugins` entry in `ci-mgmt.yaml` to
`.config/mise.toml`. To simplify the mise migrations I also refactored a
little so that we now have the following flow:

- `deleteOldMiseConfig`: Runs first and deletes the old root level
  `mise.toml` if it exists and is not the overrides file
- `migrateCimgmtOverrides`: Migrates things from `ci-mgmt.yaml` to the
  root level override file (currently just `toolVersions`).
- `migrateCimgmtToMise`: Migrates things from `ci-mgmt.yaml` to the
  `.config/mise.toml` file.
- `maintainMiseLock`: Runs last and runs `mise install` to update the
  lockfiles.

I've set the plugins version to "latest" for all the plugins so that we
can implement an upgrade workflow (e.g. via `mise upgrade`). Also, since
the plugins are added to the `.config/mise.toml` we can add specific
version overrides to the root level `mise.toml`.

## Additional things

- I've added some generic functionality for dealing with `toml` and
`yaml` files so it will be easier to maintain these going forward
- Since `mise` can automatically install `plugins`, I've also created a
backend plugin for installing the plugins
https://github.com/pulumi/vfox-pulumi
- This plugin will be automatically installed with `mise install` so
that the tools defined using the plugins can be installed.
- Added `mise_install` and `mise_env` make targets. This should make it
so that as long as you have `mise` installed, running any of the `make`
targets should automatically use the `mise` environment.

closes #1474