Add gates for backported version releases (#1808)
This pull request:

- Determines whether the tag that triggered the `publish` Workflow was
based off of the default branch (i.e. a regular release) or off a
feature branch (as would be the case for a backport release).
- Uses the `make_latest` [input to the release
Action](https://github.com/softprops/action-gh-release?tab=readme-ov-file#-customizing)
to explicitly state whether this release should be marked as `latest` on
GitHub, using whether we're on `main` or not as the logic to determine
the value
- Gates a docs build on whether we're on `main`/`master` as well.

Considerations:

- This change allows for _any_ tags on default branch to set the release
as `latest`; it doesn't actually have to be the tip of default. As long
as default is the head ref for the tag, the release will be marked as
latest. This is the case today; this code does not change that.

Fixes #1806.

- **Determine the branch name of the event's head ref. Set make_latest
to the value of whether the branch name is default or not. If we're not
pushing this tag from default, make-latest will be set to false.**
- **Do not build docs for backported releases. Add comments**