Respect noSchema in native providers (#1836)
We currently hard-code the terraform and command providers as opting-out
of schema checks.

We already expose config to let providers do this themselves, and we
respect it in bridged providers. Let's also respect it in native
providers.

This unblocks https://github.com/pulumi/pulumi-std/pull/96 because that
provider doesn't currently keep its schema in a standard place.

To be merged after:
* https://github.com/pulumi/pulumi-terraform/pull/891
* https://github.com/pulumi/pulumi-command/pull/1023