Revert "Fix duplicated env vars in native providers (#1795)"
This reverts commit 5619381e924fbac90750f41b6a27cadcd11c4c2d.

The revert had an unintended side effect of removing
`PULUMI_LOCAL_NUGET` for providers that _don't_ specify `envOverride`,
which in retrospect should have been obvious.

Let's roll the change back for the sake of unbreaking things. We can
revisit the original issue with `envOverride` later as that likely
affects a smaller number of providers.