Move back `build` target to top of Makefile for `base` (#1838)
Fixes #1825

From the Makefile manual:

> By default, the goal is the first target in the makefile (not counting
targets that start with a period). Therefore, makefiles are usually
written so that the first target is for compiling the entire program or
programs they describe.

When we introduced mise_install target on top of base/Makefile, we
replaced the build target as goal.

## Testing

Checked out `test-providers` and verified they worked as before the
`mise` change:
* [x] acme
* [x] aws
* [x] cloudflare
* [x] docker
* [x] eks
* [x] xyz