Set GOTOOLCHAIN to `auto` in Verify Release Workflows (#1853)
This pull request is a hotfix for
https://github.com/pulumi/pulumi-tls/issues/888.

In all other workflows, we allow for Mise to install the Go version but
we decided to keep the legacy implementation where we allow for the
setup-go job to install the version instead.
Setup-go shipped a change https://github.com/actions/setup-go/pull/460
where the gotoolchain is now installed using `local` not `auto` and this
keeps the verify-release workflow from installing a Go version other
than the one set in the tool version.

Update: It turns out that v1.22.x is not enough:
https://github.com/pulumi/pulumi-gcp/actions/runs/19448883749/job/55655199413#step:17:54

> Error: exec: go: go.mod requires go >= 1.23.3 (running go 1.21.13;
GOTOOLCHAIN=local)

so rather than bumping the Go version, this pull request sets the go
toolchain to `auto` so the workflow downloads it as before.