Run migrations as part of testing (#1863)
We don't currently run migrations as part of testing or CI, which means
we don't have good visibility into what the final mise.toml will look
like.

Migrations are slow, and they also aren't always possible to run due to
relevant files not being present.

For the first problem, we can modify the make target to generate and
test providers in parallel. This is a nice speedup at the cost of
slightly more confusing logs -- I've added some logging prefixes to help
alleviate that somewhat. The tradeoff is worth it.

For the second problem there are only 2 migrations which assume a
`./provider/resources.go` file is present, and we can update those to
short-circuit if it's missing.

The end result allows us to check in mise.toml and we get to see the net
effect of our migrations on ci-mgmt.yml. In particular it will allow us
to play around with these environments locally!

An "activate mise" migration is included, because the CI tests were
using ci-mgmt's toolchain instead of the provider's toolchain. That
caused us to inadvertently migrate the provider's .golangci.yaml (since
the providers are still on v1 but ci-mgmt is on v2).