Relax 'worktree clean' check for Renovate PRs (#1903)
Relevant to two internal conversations.

*
https://pulumi.slack.com/archives/C09ANN8P0N4/p1765223435588129?thread_ts=1765220623.591769&cid=C09ANN8P0N4
-- wherein we started generating a new `.gitattributes` file.
* https://pulumi.slack.com/archives/C0A2UPPKCC8/p1765054040354419 --
wherein a dependency bump made changes to `schema.json`.

In both of these cases the PR was automated but blocked because we found
the worktree to be dirty after rebuilding. But in both cases we were
only bumping dependencies, so we can assume any changes to generated
artifacts is due to upstream.

This PR relaxes the worktree-is-clean check to allow for changes on
Renovate PRs. Immediately after those changes are detected we commit
them.